Use payload karpenter if present, otherwise don't
Right now hypershift is just hardcoded to use the upstream 1.0.7 version
of Karpenter. We want to deploy our downstream version instead, but it's
not in the payload yet.

This adds logic to check the payload and use the image if it's present,
but if it's not fall back to the existing behavior. This will let us
test and eventually ship seamlessly once the image is in the payload,
but will also not break hypershift in the mean time.