Sync CRDs from downstream OpenShift
Our downstream fork is newer, so switching to it required library bumps,
and along with those type changes came some CRD changes, so this just
syncs the CRDS from vendor so types/libs/CRDs are all in sync.

This also adds an extremely rudimentary crds-sync.sh script that we can
improve if we want to do it this way. Embed won't embed files from
vendor, it explicitly ignores them, so we have to copy them,

Long term we probably need a better mechanism.