feat(autonode): Allow role/instanceprofile manipulation in autonode
Design documents intended instanceprofile/node to be adjustable by
users, but we hadn't added the plumbing to OpenShiftEc2NodeClass to do
that yet.

This adds the fields to OpenShiftEc2NodeClass and adds the sync
plumbing to the controller in the karpenter operator. Also adds tests to
enforce the functionality.

This also adjusts the default ec2nodeclass sync in the karpenter
controller to preserve any existing user-set fields that we allowed to
be set that were already there. (Previously it just overwrote them all
when it reconciled)

Signed-off-by: John Kyros <jkyros@redhat.com>

fix(autonode): Preserve user set fields in default nodeclass

Signed-off-by: John Kyros <jkyros@redhat.com>